+++ src/listener.rs
@@ -1,8 +1,11 @@
-pub(crate) use serde::{Deserialize, Serialize};
-use serde_json::{self, Result, Value};
 
-use std::io::{BufRead, BufReader, Write};
-use std::net::TcpStream;
+
+use std::error::Error;
+
+pub(crate) use serde::{Deserialize, Serialize};
+use serde_json::{self, Value};
+use tokio::io::{AsyncBufReadExt, AsyncWriteExt};
+use tokio::{io::BufReader, net::TcpStream};
 
 use crate::storage::{get, reset, set, size};
 
@@ -14,63 +17,63 @@ pub struct MemoizerMessage {
     p: Value,  // payload
 }
 
-pub fn respond(message: &str, mut stream: &TcpStream) {
-    let paylaod = format!("{}\n\r", message);
-    stream.write_all(paylaod.as_bytes());
-}
-
-// Performs the corresponding action for the given 
-// MemoizerMessage and responds its answer using the given stream.
-pub fn route(message: MemoizerMessage, mut stream: &TcpStream) {
+// Performs the corresponding action for the given
+// MemoizerMessage and returns the corresponding answer
+pub fn route(message: MemoizerMessage) -> String {
     match message.s.as_str() {
         "get" => {
             let key = message.p["k"].to_string();
             let value = get(key);
             match value {
-                None => respond("null", stream),
-                Some(v) => respond(&v, stream),
+                None => "null".to_string(),
+                Some(v) => v,
             }
         }
         "set" => {
             let key = message.p["k"].to_string();
             let value = message.p["v"].to_string();
             set(key, value);
-            respond("ok", stream)
+            "ok".to_string()
         }
         "reset" => {
             reset();
-            respond("ok", stream)
+            "ok".to_string()
         }
         "size" => {
             let size = size();
-            respond(&format!("{}", size), stream)
+            format!("{}", size)
         }
         _ => {
             println!("Received and unsupported value");
-            respond(&format!("nok: {:?}", message), stream)
+            format!("nok: {:?}", message)
         }
     }
 }
 
 // Handler that responds to the MemoizerMessage on the given line and stream.
-fn on_line_received(line: String, stream: &TcpStream) {
-    let m: Result<MemoizerMessage> = serde_json::from_str(&line);
-    match m {
-        Ok(m) => route(m, stream),
+fn on_line_received(message: serde_json::Result<MemoizerMessage>) -> String {
+    match message {
+        Ok(m) => route(m),
         Err(err) => {
             println!("Received and unsupported value");
             let error_message = format!("{:?}", err);
-            respond(&error_message, stream)
+            error_message
         }
     }
 }
 
 // Handler for a new incoming connection.
 // Will parse content as one MemoizerMessage per line.
-pub fn on_socket_accept(stream: &TcpStream) {
-    println!("Accepting incoming connection: {:?}", stream.local_addr());
-    let source = BufReader::new(stream);
-    for line in source.lines() {
-        on_line_received(line.unwrap(), stream);
-    }
+pub async fn on_socket_accept(mut stream: TcpStream) -> Result<(), Box<dyn Error>> {
+    let (reader, mut writer) = stream.split();
+    let mut lines = BufReader::new(reader).lines();
+    while let Some(line) = lines.next_line().await? {
+        let message: serde_json::Result<MemoizerMessage> = serde_json::from_str(&line);
+        let response = on_line_received(message);
+        let paylaod = format!("{}\n\r", response);
+        writer.write_all(paylaod.as_bytes())
+            .await
+            .expect("Failed to write to the socket");
+    };
+    Ok(())
 }