+++ src/main.rs
@@ -2,32 +2,27 @@
 extern crate lazy_static;
 extern crate mut_static;
 
-use crate::listener::*;
-use crate::config::*;
-use std::net::TcpListener;
-use std::thread;
+use tokio::net::TcpListener;
 
+use crate::config::*;
+use crate::listener::*;
+use std::error::Error;
+
+pub mod config;
 mod listener;
 pub mod storage;
-pub mod config;
 
-fn main() {
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn Error>> {
     let socket_address = get_socket_address();
+    let listener = TcpListener::bind(&socket_address).await?;
 
-    // Bind to socket
-    let listener = match TcpListener::bind(&socket_address) {
-        Err(_) => panic!("Failed to bind socket"),
-        Ok(listener) => listener,
-    };
-
-    println!("Server started, waiting for clients on {}", socket_address);
-
-    listener.incoming().for_each(|stream| match stream {
-        Ok(stream) => {
-            thread::spawn(move || on_socket_accept(&stream));
-        }
-        Err(err) => {
-            println!("Error: {}", err);
-        }
-    });
+    loop {
+        let (socket, _) = listener.accept().await?;
+        let _thread = tokio::spawn(async move {
+            on_socket_accept(socket)
+                .await
+                .expect("Failed to process incoming socket connection");
+        });
+    }
 }